pip-requirements.txt
git+https://github.com/ChlorideCull/DerPyBooru.git@non-standard-changes
python-telegram-bot==3.1.2